0.1基础需求

+++ src/main/java/com/company/project/service/ReplyService.java
@@ -1,9 +1,7 @@
 package com.company.project.service;
 import com.company.project.model.Reply;
 import com.company.project.core.Service;
-import com.company.project.web.model.ReplyResult;
-
-import java.math.BigDecimal;
+import com.company.project.service.model.ReplyResult;
 import java.util.List;
 
 
@@ -11,7 +9,7 @@ import java.util.List;
  * Created by CodeGenerator on 2018/11/06.
  */
 public interface ReplyService extends Service<Reply> {
-    List<ReplyResult> getReplys(BigDecimal postid);
+    List<ReplyResult> getReplys(Integer postid);
 
 
 }
